+++ example/lib/example.dart
@@ -11,7 +11,7 @@ part 'example.g.dart';
 
 @RestApi(baseUrl: 'https://5d42a6e2bc64f90014a56ca0.mockapi.io/api/v1/')
 abstract class RestClient {
-  factory RestClient(Dio dio, {String baseUrl}) = _RestClient;
+  factory RestClient(Dio dio, {String baseUrl}) = RestClientYmlp;
 
   @GET('/tasks/{id}')
   Future<ApiResult<Task?>> getNestApiResultGenericsInnerTypeNullable();
@@ -62,6 +62,11 @@ abstract class RestClient {
 
   @PUT('/tasks/{id}')
   Future<Task> updateTask(@Path() String id, @Body() Task task);
+
+  @PreventNullToAbsent()
+  @PATCH('/tasks/{id}')
+  Future<Task> updateTaskAvatar(
+      @Path() String id, @Field('avatar') String? avatar);
 
   @DELETE('/tasks/{id}')
   Future<void> deleteTask(@Path() String id);
@@ -127,7 +132,7 @@ abstract class RestClient {
   @POST('/post')
   Future<String> postFormData3({
     @Part(value: 'customfiles', contentType: 'application/json')
-        required List<File> files,
+    required List<File> files,
     @Part(fileName: 'abc.txt') required File file,
   });
 
@@ -154,7 +159,10 @@ abstract class RestClient {
   Future<String> queries(@Queries() Map<String, dynamic> queries);
 
   @GET('/enums')
-  Future<String> queryByEnum(@Query('tasks') TaskQuery query);
+  Future<String> queryByEnum(
+    @Query('tasks') TaskQuery query,
+    @Query("date") DateTime time,
+  );
 
   @GET('/get')
   Future<String> namedExample(
@@ -240,6 +248,14 @@ abstract class RestClient {
   @POST('post')
   Future<String> nestedGenericOther(
     @Body() ValueWrapper<ValueWrapper<TaskQuery>> request,
+  );
+
+  @MultiPart()
+  @POST('post/{id}/comments/{commentId}')
+  Future<String> multipartBodyWithMultiplePathParameter(
+    @Path("id") String id,
+    @Path("commentId") String commentId,
+    @Part() Map<String, dynamic> body,
   );
 }